+++ requirements_base.txt
@@ -4,3 +4,4 @@ argparse==1.2.1
 noise==1.2.2
 nose==1.3.7
 numpy==1.9.2
+pygdal==1.11.2.1
