+++ setup.py
@@ -1,16 +1,5 @@
 from setuptools import setup, Extension, Command
-# from pip.req import parse_requirements
-# import pip.download
-
-# parse_requirements() returns generator of pip.req.InstallRequirement objects
-# install_reqs = parse_requirements("requirements2.txt",
-# session=pip.download.PipSession())
-
-# reqs is a list of requirement
-# e.g. ['django==1.5.1', 'mezzanine==1.4.6']
-# reqs = [str(ir.req) for ir in install_reqs]
-
-execfile('worldengine/version.py')
+from worldengine.version import __version__
 
 config = {
     'name': 'worldengine',
