+++ worldengine/draw.py
@@ -1,6 +1,6 @@
 from PIL import Image
 
-from worldengine.drawing_functions import draw_ancientmap, \
+from drawing_functions import draw_ancientmap, \
     draw_rivers_on_image, gradient
 
 # -------------
