+++ worldengine/simulations/ErosionSimulation.py
@@ -1,6 +1,6 @@
 import math
 import numpy
-import worldengine.a_star
+from utilities import astar
 
 # Direction
 NORTH = [0, -1]
@@ -230,7 +230,7 @@ class ErosionSimulation(object):
             is_wrapped, lower_elevation = self.findLowerElevation(
                 current_location, world)
             if lower_elevation and not is_wrapped:
-                lower_path = worldengine.a_star.PathFinder().find(
+                lower_path = astar.PathFinder().find(
                     world.elevation['data'], current_location, lower_elevation)
                 if lower_path:
                     path += lower_path
@@ -273,7 +273,7 @@ class ErosionSimulation(object):
                             current_location, lower_elevation))
 
                 # find our way to the edge
-                edge_path = worldengine.a_star.PathFinder().find(
+                edge_path = astara_star.PathFinder().find(
                     world.elevation['data'], [cx, cy], [lx, ly])
                 if not edge_path:
                     # can't find another other path, make it a lake
@@ -284,7 +284,7 @@ class ErosionSimulation(object):
                 current_location = path[-1]
 
                 # find our way to lowest position original found
-                lower_path = worldengine.a_star.PathFinder().find(
+                lower_path = astar.PathFinder().find(
                     world.elevation['data'], current_location, lower_elevation)
                 path += lower_path
                 current_location = path[-1]
